Makefile: fix arguments to go, not matching what go expects
This caused gcc to receive the actual argument "''" or for example
"'-Wl,-O1,--sort-common,--as-needed,-z,relro,-z,now'" and then error
with
gcc: error: '-Wl,-O1,--sort-common,--as-needed,-z,relro,-z,now': No such file or directory

Furthermore, the trimpath arguments tried to trim an actual path
resembling "'$PWD'", which did nothing as the literal single quotes were
not valid in this context.

The former issue means it is down to the go compiler to interpret
-ldflags properly in order to break up the (multiple) flag values and
correctly parse the extldflags, possible spaces included.

In the latter case, trimpath only accepts one argument and the
shellscript already correctly double-quotes this in order to pass it as
one argv[3] to the compiler:
{ "go", "build", "-gcflags", "all=-trimpath=/path/to/pwd" }

And from there the compiler should be able to interpret paths with
spaces on its own.